Ready for another reinstall

+++ arch-install.sh
@@ -17,9 +17,13 @@ setupefi(){
 if $(prompt "Is EFI already setup?"); then
     echo ""
     mount /dev/sda3 /mnt
+    mkdir /mnt/boot
+    mkdir /mnt/boot/efi
+    mount /dev/sda1 /mnt/boot/efi
     return
 
 else
+    echo ""
     mkfs.vfat /dev/sda1
     mount /dev/sda3 /mnt
     mkdir /mnt/boot
@@ -31,32 +35,43 @@ fi
 
 install_system(){
 	#fucking goddamn network
-	log "Time to set my clock and partition the drive!"
-	
-    timedatectl set-ntp true
-	#Let user set up the parts
-	cfdisk /dev/sda
-    #Make Swap and format the parts
-	mkswap /dev/sda2
-    swapon /dev/sda2
-	mkfs.ext4 /dev/sda3
 
-    setupefi
-	
-    #Install the base packages
-    pacstrap /mnt base base-devel acpi refind-efi htop networkmanager openssh git
-
-    #Gen the Fstab
-    genfstab -p -U /mnt >> /mnt/etc/fstab
-
-    #copy part 2 to /mnt
-    mkdir /mnt/installer
-    cp arch-install-chroot.sh /mnt/installer/
-    #Chroot and run other part of installer
-    arch-chroot /mnt sh /installer/arch-install-chroot.sh
+    wget -q --tries=10 --timeout=20 --spider http://google.com
+    if [[ $? -eq 0 ]]; then
+        log "We're Online Baby!"
+        log "Time to set my clock and partition the drive!"
     
-    #Unmount that shit
-    umount /mnt
+        timedatectl set-ntp true
+        #Let user set up the parts
+        cfdisk /dev/sda
+        #Make Swap and format the parts
+        mkswap /dev/sda2
+        swapon /dev/sda2
+        mkfs.ext4 /dev/sda3
+
+        setupefi
+        
+        #Install the base packages
+        pacstrap /mnt base base-devel acpi refind-efi htop networkmanager openssh git dialog
+
+        #Gen the Fstab
+        genfstab -p -U /mnt >> /mnt/etc/fstab
+
+        #copy part 2 to /mnt
+        mkdir /mnt/installer
+        cp arch-install-chroot.sh /mnt/installer/
+        #Chroot and run other part of installer
+        arch-chroot /mnt sh /installer/arch-install-chroot.sh
+        
+        #Unmount that shit
+        umount /mnt
+        log "Be sure to check out your fstab before rebooting!"
+    else
+        log "Actually connect to the internet and try again asswipe!"
+        return
+    fi
+
+	
 }
 
 
@@ -77,7 +92,6 @@ echo "-------------------------------------------------------------------------"
 if $(prompt "Do you agree to this scheme?"); then
     echo ""
     install_system
-    log "Be sure to check out your fstab before rebooting!"
 
 else
     echo ""
